refactor: 整理model文件
