+++ lib/model/status/attribute_model.dart
